local opy update

+++ kwave/utils/kwave_array.py
@@ -10,7 +10,7 @@ from math import ceil
 
 from kwave.utils.conversion import tol_star
 from kwave.utils.interp import get_delta_bli
-from kwave.utils.mapgen import trim_cart_points, make_cart_rect, make_cart_arc
+from kwave.utils.mapgen import trim_cart_points, make_cart_rect, make_cart_arc, make_cart_spherical_segment
 from kwave.utils.math import sinc, get_affine_matrix
 from kwave.utils.matlab import matlab_assign, matlab_mask, matlab_find
 
@@ -139,7 +139,8 @@ class kWaveArray(object):
                 element_number=el_ind + 1,
                 type='annulus',
                 dim=2,
-                position=array(position, dtype=np.uint8),
+                #position=array(position, dtype=np.uint8), # bad cast.
+                position=array(position),
                 radius_of_curvature=radius,
                 inner_diameter=diameters[el_ind][0],
                 outer_diameter=diameters[el_ind][1],
